Scene1 generation working

+++ SourceCode/SimpleRayTracer/Color.h
@@ -2,19 +2,19 @@
 
 class Color {
 public:
-	uint8_t r, g, b, a;
+	int r, g, b, a;
 
 	Color() : r(0), g(0), b(0), a(1)
 	{
 
 	}
 
-	Color(uint8_t r, uint8_t g, uint8_t  b) : r(r), g(g), b(b), a(1)
+	Color(int r, int g, int  b) : r(r), g(g), b(b), a(1)
 	{
 
 	}
 
-	Color(uint8_t r,uint8_t g,uint8_t  b,uint8_t a) : r(r), g(g), b(b), a(a)
+	Color(int r,int g,int b,int a) : r(r), g(g), b(b), a(a)
 	{
 
 	}
